+++ s4e-backend/src/main/java/pl/cyfronet/s4e/data/repository/query/QueryText.java
@@ -4,22 +4,24 @@ import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
 
+import static pl.cyfronet.s4e.search.SearchQueryParams.*;
+
 public class QueryText extends QueryDecorator {
     Map<String, String> paramsToDBField;
 
     public QueryText(QueryBuilder queryBuilder) {
         super(queryBuilder);
         paramsToDBField = new LinkedHashMap<>();
-        paramsToDBField.put("satellitePlatform", "spacecraft");
-        paramsToDBField.put("productType", "product_type");
-        paramsToDBField.put("polarisation", "polarisation");
-        paramsToDBField.put("processingLevel", "processing_level");
-        paramsToDBField.put("sensorMode", "sensor_mode");
+        paramsToDBField.put(SATELLITE_PLATFORM, "spacecraft");
+        paramsToDBField.put(PRODUCT_TYPE, "product_type");
+        paramsToDBField.put(POLARISATION, "polarisation");
+        paramsToDBField.put(PROCESSING_LEVEL, "processing_level");
+        paramsToDBField.put(SENSOR_MODE, "sensor_mode");
         paramsToDBField.put("collection", "collection");
         paramsToDBField.put("timeliness", "timeliness");
         paramsToDBField.put("instrument", "instrument");
         paramsToDBField.put("productLevel", "product_level");
-        paramsToDBField.put("relativeOrbitNumber", "relative_orbit_number");
+        paramsToDBField.put(RELATIVE_ORBIT_NUMBER, "relative_orbit_number");
         paramsToDBField.put("absoluteOrbitNumber", "absolute_orbit_number");
     }