+++ s4e-backend/src/main/java/pl/cyfronet/s4e/search/SearchQueryParams.java
@@ -1,0 +1,21 @@
+package pl.cyfronet.s4e.search;
+
+public class SearchQueryParams {
+    public static final String SORT_BY = "sortBy";
+    public static final String ORDER = "order";
+
+    public static final String SENSING_FROM = "sensingFrom";
+    public static final String SENSING_TO = "sensingTo";
+
+    public static final String INGESTION_FROM = "ingestionFrom";
+    public static final String INGESTION_TO = "ingestionTo";
+
+    public static final String SATELLITE_PLATFORM = "satellitePlatform";
+    public static final String PRODUCT_TYPE = "productType";
+    public static final String PROCESSING_LEVEL = "processingLevel";
+    public static final String POLARISATION = "polarisation";
+    public static final String SENSOR_MODE = "sensorMode";
+    public static final String RELATIVE_ORBIT_NUMBER = "relativeOrbitNumber";
+
+    public static final String CLOUD_COVER = "cloudCover";
+}