+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The most recent changes are on top, in each type of changes category.
 
 ### Added
 
+- Add Sentinel search config endpoint [#485](https://github.com/cyfronet-fid/sat4envi/issues/485)
 - Preset gs-gateway configurations per dataset
 - Edit institution [#494](https://github.com/cyfronet-fid/sat4envi/issues/494)
 - GeoServer Gateway [#479](https://github.com/cyfronet-fid/sat4envi/issues/479)